requirements.txt

asgiref>=3.5.2,<4
setuptools
autopep8==1.5.7
dj-database-url==0.5.0
Django==4.2.13
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
requests
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
paho-mqtt
channels
channels_redis
djangorestframework
daphne
django-import-export
celery
celery[redis]
django-redis
django-celery-beat
django-import-export
django-multiselectfield
django-redis
django-timezone-field
djangorestframework
mysqlclient
djangorestframework-simplejwt
mysqlclient
requests








